(feat): reorganize dirs
