+++ anndata/experimental/read_backed/__init__.py
@@ -1,1 +1,2 @@
-from .read_backed import read_backed, LazyCategoricalArray, LazyMaskedArray
+from .read_backed import read_backed
+from .lazy_arrays import LazyCategoricalArray, LazyMaskedArray
